AdminResultaten hernoemd naar AdminScores + Allerlei aanpassingen aan activities & Design + logo.png hernoemd naar afbeeldingSplashScreen

+++ app/src/main/java/be/thomasmore/androidproject/Conditie1Oefening3.java
@@ -84,7 +84,7 @@ public class Conditie1Oefening3 extends AppCompatActivity {
         if (contextzin.getText().equals(contextzinnen.get(1))) {
             Bundle bundle = new Bundle();
             bundle.putLong("studentID", studentID);
-            bundle.putLong("index", index);
+            bundle.putInt("index", index);
             bundle.putLongArray("woordIDs", woordIDs);
             Intent intent = new Intent(this, Conditie1Oefening4.class);
             intent.putExtras(bundle);
